+++ src/collectl2pcp/load.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013 Red Hat Inc.
+ * Copyright (c) 2013 Red Hat.
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the
@@ -13,7 +13,6 @@
  *
  * Handler for load average
  * "load 1.07 1.18 1.30 3/658 4944"
- *
  */
 
 #include "metrics.h"
@@ -24,7 +23,7 @@ loadavg_handler(handler_t *h, fields_t *f)
     pmInDom indom = pmInDom_build(LINUX_DOMAIN, LOADAVG_INDOM);
     if (f->nfields < 4)
     	return -1;
-    if (!isdigit(f->fields[1][0]) || !isdigit(f->fields[2][0]) || !isdigit(f->fields[3][0])) {
+    if (!isdigit((int)(f->fields[1][0])) || !isdigit((int)(f->fields[2][0])) || !isdigit((int)(f->fields[3][0]))) {
     	fprintf(stderr, "Warning: corrupted value for kernel load average: \"%s %s %s\" ignored.\n",
 	    f->fields[1], f->fields[2], f->fields[3]);
 	return -1;
